grsec: Apply grsecurity-2.9.1-3.2.42-201304082214.patch
commit 37fca32e3fda5f755fa583d2c6e60856fa0f88e8
Merge: 8d5c7f0 356f77d
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Apr 8 21:31:20 2013 -0400

    Merge branch 'pax-stable2' into grsec-stable2

    Conflicts:
    	arch/sparc/kernel/us3_cpufreq.c
    	fs/proc/base.c

commit 356f77d7ff0e77de3427b9c2d937f5aaa83b7337
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Apr 8 21:20:54 2013 -0400

    Update to pax-linux-3.2.42-test101.patch:
    - fixed some attribute leakage into userland headers, patch by Mathias Krause
    - fixed some of the access_*_vm related breakage that trigger size overflows, reported by Hunger
    - fixed ia64/ppc/sparc compilation by spender
    - improved the STRUCTLEAK gcc plugin to cover a few more cases (credit to stef for bugreport)

 arch/ia64/include/asm/uaccess.h          |    2 -
 arch/powerpc/include/asm/uaccess.h       |    2 -
 arch/powerpc/platforms/cell/spufs/file.c |    4 +-
 arch/sparc/include/asm/uaccess.h         |    7 ---
 arch/sparc/kernel/prom_common.c          |    2 +-
 arch/sparc/kernel/us3_cpufreq.c          |   69 ++++++++---------------------
 fs/proc/base.c                           |    4 +-
 fs/sysfs/bin.c                           |    6 +-
 include/linux/mm.h                       |   16 +++---
 mm/memory.c                              |   24 +++++-----
 mm/nommu.c                               |   10 ++--
 scripts/headers_install.pl               |    1 +
 tools/gcc/structleak_plugin.c            |   15 +++---
 13 files changed, 61 insertions(+), 101 deletions(-)

Signature-tree: 9ae6e0239d79c5116595e40e03b44e47ce6afb81

+++ fs/sysfs/bin.c
@@ -233,13 +233,13 @@ static int bin_page_mkwrite(struct vm_area_struct *vma, struct vm_fault *vmf)
 	return ret;
 }
 
-static int bin_access(struct vm_area_struct *vma, unsigned long addr,
-		  void *buf, int len, int write)
+static ssize_t bin_access(struct vm_area_struct *vma, unsigned long addr,
+		  void *buf, size_t len, int write)
 {
 	struct file *file = vma->vm_file;
 	struct bin_buffer *bb = file->private_data;
 	struct sysfs_dirent *attr_sd = file->f_path.dentry->d_fsdata;
-	int ret;
+	ssize_t ret;
 
 	if (!bb->vm_ops)
 		return -EINVAL;
